Replace “@” to “♥♥” and replace “=” to “▒”
Improve on the user interface. Replace the food symbol with double heart symbol (“♥♥”) so that it is easier for the snake to reach the food and the wall with “▒” symbol.

Another criteria is added in the if statement to check whether the snake could consume the other part of the food.

Another if statement to check if the food is spawned at the last column in the screen. If true, the food’s column in the position would be reduced by 1, so that the other part of the food would be spawned at the last column in the screen.